[rtlibc] remove libc_signal.h and libc_fdset.h
