+++ components/libc/compilers/common/none-gcc/sys/types.h
@@ -8,8 +8,8 @@
  * 2020-09-05     Meco Man     fix bugs
  * 2020-12-16     Meco Man     add useconds_t
  */
-#ifndef __TYPES_H__
-#define __TYPES_H__
+#ifndef __SYS_TYPES_H__
+#define __SYS_TYPES_H__
 
 #include <stdint.h>
