+++ components/libc/pthreads/mqueue.c
@@ -8,6 +8,7 @@
  */
 
 #include <string.h>
+#include <sys/signal.h>
 #include "mqueue.h"
 #include "pthread_internal.h"
