+++ components/libc/signal/posix_signal.c
@@ -7,6 +7,7 @@
  * Date           Author       Notes
  * 2017/10/1      Bernard      The first version
  */
+
 #include <rthw.h>
 #include <rtthread.h>
 
@@ -14,6 +15,7 @@
 #include <sys/errno.h>
 
 #include "posix_signal.h"
+
 #define sig_valid(sig_no)   (sig_no >= 0 && sig_no < RT_SIG_MAX)
 
 void (*signal(int sig, void (*func)(int))) (int)
@@ -79,13 +81,10 @@ int sigtimedwait(const sigset_t *set, siginfo_t *info,
     int ret  = 0;
     int tick = RT_WAITING_FOREVER;
 
-#ifdef RT_USING_PTHREADS
     if (timeout)
     {
-        extern int clock_time_to_tick(const struct timespec *time);
         tick = clock_time_to_tick(timeout);
     }
-#endif
 
     ret = rt_signal_wait(set, info, tick);
     if (ret == 0) return 0;
