+++ include/rtlibc.h
@@ -13,11 +13,8 @@
 
 /* definitions for libc if toolchain has no these definitions */
 #include "libc/libc_stat.h"
-
 #include "libc/libc_fcntl.h"
 #include "libc/libc_dirent.h"
-#include "libc/libc_signal.h"
-#include "libc/libc_fdset.h"
 
 #ifndef RT_USING_LIBC
 #if defined(__ARMCC_VERSION) || defined(__IAR_SYSTEMS_ICC__)
